refactor: rename Footer-styling file
